Files Update
I made some edits to _script.js._ It's still the main js file in the bunch.

Added _profiles.js_ to hold the profile objects made by users.

Also added an _itemDescription.txt_ file to hold... you get the idea. I'll need to come up with a good way of organizing the descriptions so that they can be easily accessed by the other files. It may be best to simply use an excel file type.